requirements.txt
# Core ML (3.13-compatible)
numpy>=2.1,<2.3
pandas>=2.2.3,<2.4
scikit-learn>=1.6,<1.8

# API
flask>=3.0,<3.1

# Database
sqlalchemy>=2.0.23,<2.1

# Tracking
mlflow>=3.5.1,<3.6

# Utilities
pyyaml>=6.0.2,<7

xgboost>=3.1,<3.2
